extract components out of CommentItem
