+++ client/scenes/document/scenes/Version/components/CommentContainer/CommentContainer.scss
@@ -158,7 +158,7 @@
 }
 
 .comment-item__collapse-btn {
-  padding-bottom: 1.3rem;
+  padding-bottom: 1.1rem;
   border-bottom: 1px solid;
 }
