+++ client/scenes/document/scenes/Version/components/CommentContainer/index.js
@@ -1,0 +1,7 @@
+export { default as ActionableIssueTag } from "./ActionableIssueTag";
+export { default as ActionBar } from "./ActionBar";
+export { default as CommentContainer } from "./CommentContainer";
+export { default as CommentItem } from "./CommentItem";
+export { default as EditCommentModal } from "./EditCommentModal";
+export { default as MainComment } from "./MainComment";
+export { default as Replies } from "./Replies";
