+++ client/scenes/document/scenes/Version/components/index.js
@@ -7,12 +7,11 @@ export { default as VersionIssues } from "./VersionIssues";
 export { default as SidebarComments } from "./SidebarComments";
 export { default as SidebarTableOfContents } from "./SidebarTableOfContents";
 export { default as VersionScorecard } from "./VersionScorecard";
-export { default as CommentItem } from "./CommentItem/CommentItem";
-export { default as CommentModal } from "./CommentItem/CommentModal";
-export { default as CommentBox } from "./CommentBox";
+export { default as CommentItem } from "./CommentContainer/CommentContainer";
+export { default as CommentBox } from "./CommentBox/CommentBox";
 export {
   default as CommentBoxWithTagField
 } from "./CommentBox/CommentBoxWithTagField";
-export { default as EditCommentModal } from "./CommentItem/EditCommentModal";
+export { default as EditCommentModal } from "./CommentContainer/EditCommentModal";
 export { default as Countdown } from "./Countdown";
 export { default as ContentEditingContainer } from "./ContentEditingContainer";
