+++ package.json
@@ -43,6 +43,7 @@
     "express": "^4.14.1",
     "express-session": "^1.15.1",
     "formsy-react": "^1.1.4",
+    "get-urls": "^8.0.0",
     "github-markdown-css": "^2.10.0",
     "halogenium": "^2.2.3",
     "heroku-ssl-redirect": "0.0.4",
